move Map into RightPane folder
