+++ src/components/RightPane/Map/MapTabsAndSearchBar.js
