+++ src/components/RightPane/Map/static/buildingCatalogue.js
